Styling #6
Updated the orders.html template: Added the class scrollable-container to the orders container.
Added CSS: Defined the scrollable-container class to make it scrollable with a fixed height.
This will ensure that the orders are displayed in a scrollable container, making it easier to manage and view all the orders on the page.